A bit of creative writing that hits all the right buttons for someone like me. 🙂
I'm still here, alive and spying on the game's progress up the charts on Steam. I'd like you all to know I left on good terms with everyone to the best of my knowledge and have no regrets or gripes except that I couldn't quite see the community through to go-live. For all time's sake.  🙂